requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Twilio Integration
twilio==8.10.0

# Google Cloud Services
google-cloud-speech==2.22.0
google-cloud-texttospeech==2.16.3
google-auth==2.23.4

# HTTP Clients
aiohttp==3.9.0
requests==2.31.0

# File Handling
aiofiles==23.2.1

# Date Parsing
python-dateutil==2.8.2

# Environment Management
python-dotenv==1.0.0

# Development Tools (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.10.0
flake8==6.1.0